Harden integer conversions in generated SSZ code (#217)
* Harden integer conversions in generated SSZ code

Sizes and limits resolve as uint64 spec values; the generated code now
keeps each in its proper domain instead of converting blindly:

- Size-domain expressions are rejected above the platform integer range
  right after resolution (ErrPlatformOverflow), making every int
  conversion at their use sites exact.
- List-limit checks compare in uint64 against the raw spec value, so a
  limit beyond the int range never truncates. Where a limit feeds an int
  capacity argument it clamps to math.MaxInt instead of wrapping
  negative, with uint64-range literals handled at generation time (a
  literal limit above the int range previously did not even compile).
- Offset and length handling in the decoders sticks to plain int; the
  uint32 round-trips through sszLen and maxOffset are gone.
- Offset writes guard the 4-byte SSZ offset range explicitly: an object
  growing past 2^32-1 bytes returns ErrOffsetOverflowFn (new sszutils
  error constructor) instead of writing truncated offsets.

Generated output is behaviorally identical to the reflection engine
(differential fuzzing, 14.9M iterations clean) and drops the structural
incorrect-integer-conversion findings from static analysis.

* update htr-caching ssz code

* Match the reflection engine to the conversion-safety doctrine

The reflection path bounded sizes at the uint32 range structurally: the
descriptor Size/Len/BitSize/MinSize fields were uint32, getSszValueSize
returned uint32, and dynssz-size resolution rejected values above
math.MaxUint32. The generated code allows any size up to the platform
integer range as long as no 4-byte offset overflows, so the engines
disagreed on sizes in between — and reflection marshal wrote silently
truncated offsets past 4 GiB.

- Descriptor Size/Len/BitSize/MinSize (and DynFieldDescriptor's
  HeaderOffset) widen to int64; size products and container sums are
  bounded to the platform integer range with wrap-safe checks.
- dynssz-size resolution and size annotations reject values above
  math.MaxInt with ErrPlatformOverflow, matching the generated guard;
  static ssz-size literals parse into the full range.
- getSszValueSize and ReflectionCtx.SizeSSZ return int64.
- Reflection marshal writes offsets through checked helpers: a value
  past the 4-byte SSZ offset range returns ErrOffsetOverflowFn instead
  of truncating, in the container, vector, and list paths.

Sizes past the former uint32 bound now behave identically in both
engines (differential fuzzing clean); tests asserting the old bound are
updated to the platform-range doctrine, including a regression test that
a 2^32-byte vector type is accepted.

* Harden generated-code portability and 32-bit offset validation

- validate list start offsets in the unsigned domain: on 32-bit targets
  int(uint32) wraps offsets past 2^31 negative, slipping malformed input
  through the validity checks as an empty list (buffer + stream paths)
- clamp reflection list/bitlist read caps to the platform int range
  instead of rejecting large limits; divide bitlist byte caps in uint64
  before capping
- type limit literals past the portable int range as uint64 in error
  constructor arguments so the output compiles on all targets
- replace bare min/max builtins and int(min(...)) capacity clamps in
  generated code with sszutils.Min/Max/CapToInt; a target package can
  shadow the builtins with its own functions
- drop max(x, 0) around provably non-negative counts
- ride int(ctx.exprs[N]) conversions in the streaming encoder with
  analyzer suppressions (#nosec G115 + lgtm), emitted at the build sites

* fix unnecessary int64 conversions of widened descriptor fields

* Emit size arithmetic in the uint64 domain

- static size variables are uint64 throughout: the streaming encoder sums
  them straight into its offset accumulator, buffer/decode generators bind
  them once as int where the arithmetic is int (totalSize, elemSizeN —
  bind names carry the size variable's suffix so nested element scopes
  cannot shadow an outer bind that is referenced later)
- length checks compare through uintCmpExpr: literals in the portable int
  range as plain int constants, resolved expressions in uint64
- error constructors take the raw uint64 expression instead of an int cast
- decoder container offset bounds compare in uint32 (offsets are 4-byte);
  a longer region clamps via sszutils.Min since no offset can address it
- the encoderCtx exprs array is omitted when no expressions exist and
  maxOffset is only emitted with two or more dynamic fields
- replace emitted min/max builtins with sszutils.Min/Max and drop max(x, 0)
  around provably non-negative counts

* Read container offsets as uint32 and keep buffer lengths uint64

- buffer container offsets stay uint32 from the binary read through the
  bound checks and region slicing; the upper bound clamps to the 4-byte
  offset range via sszutils.Min, mirroring the streaming decoder
- buflen and the static size sum compare in uint64: an int-converted sum
  of resolved sizes could wrap negative and pass the EOF check on a
  short buffer
- exproffset accumulates the unsigned size variables directly, removing
  the int conversions from the expression-offset slice bounds

* make the value-size overflow expectation platform-aware

* align 32-bit test expectations with the branch behavior

- the large-object rejection now fires at descriptor build with the
  platform integer range message; accept both error vintages
- sizes past the platform range are rejected on 32-bit, valid on 64-bit
- limits past the platform range clamp the read caps and decode instead
  of erroring (matching generated code), on every platform

* size lists with the native length

The list branch of the reflection size walk truncated the element count
to uint32, wrapping the size of 4GiB+ lists (marshal then failed its
final length check where the generated code encodes correctly) and
comparing the wrapped count against ssz-max.

* cover the offset overflow guards

Lying delegated sizers claim multi-GiB element sizes at zero memory
cost, driving every checked offset write past the 4-byte range on the
non-seekable paths; a position-inflating encoder stub does the same for
the seekable patches. A negative delegated size exercises the entry
point guards, and the offset-overflow constructor joins the error table.

* exercise uncovered generator branches with fixture types

New fixture types drive generation through branches no other type
reached: bulk fixed-byte vectors and lists, limits past the int32
range, spec-sized strings, bit-size and duplicate-hinted shapes,
wrapper-borne static sizes, union lists and compound region minimums,
custom fields in every delegation flavor (stream-only, buffer-only,
statically sized fastssz), grouped offset tables past the bulk-padding
threshold, and cross-flavored view children (buffer-only and
streaming-only view methods plus a static view child). Payloads with
golden roots run the generated code through the differential matrix;
the atknest holder keeps the static batch invariant.

* accept the 32-bit verdicts in the offset overflow tests

The 32-bit size walk rejects the claimed totals with the platform range
error before any offset write runs, and encoder positions past the
32-bit range cannot exist there at all.

+++ reflection/treeroot.go
@@ -356,7 +356,7 @@ func (ctx *ReflectionCtx) buildRootFromLargeUint(sourceType *ssztypes.TypeDescri
 		sourceValue = sourceValPtr.Elem()
 	}
 
-	sourceLen := uint32(sourceValue.Len())
+	sourceLen := int64(sourceValue.Len())
 	expectedLen := sourceType.Size / sourceType.ElemDesc.Size
 	if sourceLen > expectedLen {
 		return sszutils.ErrVectorLengthFn(sourceLen, expectedLen)
@@ -369,7 +369,7 @@ func (ctx *ReflectionCtx) buildRootFromLargeUint(sourceType *ssztypes.TypeDescri
 	isUint64 := sourceType.ElemDesc.Kind == reflect.Uint64
 	if isUint64 {
 		for i := 0; i < int(sourceType.Size/8); i++ {
-			if uint32(i) < sourceLen {
+			if int64(i) < sourceLen {
 				hh.AppendUint64(sourceValue.Index(i).Uint())
 			} else {
 				hh.AppendUint64(0)
@@ -378,10 +378,10 @@ func (ctx *ReflectionCtx) buildRootFromLargeUint(sourceType *ssztypes.TypeDescri
 	} else {
 		b := sourceValue.Bytes()
 		hh.Append(b)
-		// int64 keeps the widening conversion from the uint32 size safe on every
-		// platform (a large-uint width is only ever 16 or 32, but the size is a
-		// spec-derived uint32 the compiler cannot bound).
-		if pad := int64(sourceType.Size) - int64(len(b)); pad > 0 {
+		// The subtraction runs in int64 so an over-long value yields a negative
+		// pad instead of wrapping (a large-uint width is only ever 16 or 32,
+		// but the size is spec-derived and the compiler cannot bound it).
+		if pad := sourceType.Size - int64(len(b)); pad > 0 {
 			hh.Append(sszutils.ZeroBytes()[:pad])
 		}
 	}
@@ -627,7 +627,7 @@ func (ctx *ReflectionCtx) buildRootFromUnion(sourceType *ssztypes.TypeDescriptor
 //   - Byte arrays use PutBytes for efficient chunk-based hashing
 //   - Arrays with max size hints include length mixing for proper limits
 func (ctx *ReflectionCtx) buildRootFromVector(sourceType *ssztypes.TypeDescriptor, sourceValue reflect.Value, hh sszutils.HashWalker, depth reflectionDepth) error {
-	vecLen := int64(sourceType.Len)
+	vecLen := sourceType.Len
 	if vecLen > math.MaxInt {
 		return sszutils.ErrPlatformOverflowFn("vector length", sourceType.Len)
 	}
@@ -635,7 +635,7 @@ func (ctx *ReflectionCtx) buildRootFromVector(sourceType *ssztypes.TypeDescripto
 	hashIndex := hh.StartTree(sszutils.TreeTypeBinary)
 
 	sliceLen := sourceValue.Len()
-	if uint32(sliceLen) > sourceType.Len {
+	if int64(sliceLen) > sourceType.Len {
 		if sourceType.Kind == reflect.Array {
 			sliceLen = int(vecLen)
 		} else {
@@ -644,7 +644,7 @@ func (ctx *ReflectionCtx) buildRootFromVector(sourceType *ssztypes.TypeDescripto
 	}
 
 	appendZero := 0
-	if uint32(sliceLen) < sourceType.Len {
+	if int64(sliceLen) < sourceType.Len {
 		appendZero = int(vecLen) - sliceLen
 	}
